Refactor `old_string` and `new_string` option definitions.
Moved `old_string` and `new_string` option definitions from `name/options.py` to `shared_options.py` for better code organization and reuse. Updated `name/cli.py` to import these options from `shared_options`.
