+++ foundrytools_cli_2/cli/name/cli.py
@@ -11,13 +11,11 @@ from foundrytools_cli_2.cli.name.options import (
     name_id,
     name_ids,
     name_string,
-    new_string,
-    old_string,
     platform_id,
     skip_name_ids,
     win_or_mac_platform_id,
 )
-from foundrytools_cli_2.cli.shared_options import base_options
+from foundrytools_cli_2.cli.shared_options import base_options, new_string, old_string
 from foundrytools_cli_2.cli.task_runner import TaskRunner
 
 cli = click.Group(help="Utilities for editing the ``name`` table.")
